Make nixops-tarball an input to nixops overlay.
The problem with the previous approach was that importing from the
NixOps derivation (to get its tarball) during the build phase of a
Hydra job is illegal; but doing it during evaluation is, of course,
just fine. So we build the nixops tarball in release.nix and then
override it in the package set that we use to create the Hydra jobs;
then it's just a regular input that we can use in the nixops
derivation.

When evaluating the overlay "normally" (i.e., outside of the Hydra
jobset), we just import directly from the derivation, which should
work just fine.